+++ RPGgame/Portfolio/Pages/Posts/Edit.cshtml.cs
@@ -23,7 +23,7 @@ namespace Portfolio.Pages.Posts
         [BindProperty]
         public Post Post { get; set; }
 
-        public async Task<IActionResult> OnGetAsync(int? id)
+        public async Task<IActionResult> OnGetAsync(long? id)
         {
             if (id == null)
             {
